--- cli/lameduck/run.go
@@ -1,12 +1,0 @@
-package lameduck
-
-import (
-	"github.com/synadia-io/nex/cli/globals"
-)
-
-func (l LameDuckOptions) Run(cfg globals.Globals) error {
-	if cfg.Check {
-		return l.Table()
-	}
-	return nil
-}
