--- cli/lameduck/validate.go
@@ -1,5 +1,0 @@
-package lameduck
-
-func (l LameDuckOptions) Validate() error {
-	return nil
-}
