--- cli/monitor/event/event.go
@@ -1,3 +1,0 @@
-package event
-
-type EventCmd struct{}
