--- cli/monitor/event/table.go
@@ -1,5 +1,0 @@
-package event
-
-func (e EventCmd) Table() error {
-	return nil
-}
