--- cli/monitor/event/validate.go
@@ -1,5 +1,0 @@
-package event
-
-func (e EventCmd) Validate() error {
-	return nil
-}
