--- cli/monitor/log/log.go
@@ -1,3 +1,0 @@
-package log
-
-type LogCmd struct{}
