--- cli/monitor/log/run.go
@@ -1,9 +1,0 @@
-package log
-
-import (
-	"github.com/synadia-io/nex/cli/globals"
-)
-
-func (l LogCmd) Run(cfg globals.Globals) error {
-	return nil
-}
