--- cli/monitor/log/table.go
@@ -1,5 +1,0 @@
-package log
-
-func (l LogCmd) Table() error {
-	return nil
-}
