--- cli/monitor/log/validate.go
@@ -1,5 +1,0 @@
-package log
-
-func (l LogCmd) Validate() error {
-	return nil
-}
