--- cli/monitor/run.go
@@ -1,12 +1,0 @@
-package monitor
-
-import (
-	"github.com/synadia-io/nex/cli/globals"
-)
-
-func (m MonitorOptions) Run(cfg globals.Globals) error {
-	if cfg.Check {
-		return m.Table()
-	}
-	return nil
-}
