+++ cli/monitor/table.go
@@ -7,7 +7,7 @@ import (
 	"github.com/jedib0t/go-pretty/v6/text"
 )
 
-func (m MonitorOptions) Table() error {
+func (e EventCmd) Table() error {
 	tw := table.NewWriter()
 	tw.SetStyle(table.StyleRounded)
 
@@ -17,9 +17,27 @@ func (m MonitorOptions) Table() error {
 	tw.SetTitle("Monitor Configuration")
 	tw.AppendHeader(table.Row{"Field", "Value", "Type"})
 	tw.AppendRows([]table.Row{
-		{"Node ID", m.NodeId, "string"},
-		{"Workload ID", m.WorkloadId, "string"},
-		{"Level", m.Level, "string"},
+		{"Node ID", e.sharedMonitorOptions.NodeId, "string"},
+		{"Workload ID", e.sharedMonitorOptions.WorkloadId, "string"},
+		{"Level", e.sharedMonitorOptions.Level, "string"},
+	})
+	fmt.Println(tw.Render())
+	return nil
+}
+
+func (l LogCmd) Table() error {
+	tw := table.NewWriter()
+	tw.SetStyle(table.StyleRounded)
+
+	tw.Style().Title.Align = text.AlignCenter
+	tw.Style().Format.Header = text.FormatDefault
+
+	tw.SetTitle("Monitor Configuration")
+	tw.AppendHeader(table.Row{"Field", "Value", "Type"})
+	tw.AppendRows([]table.Row{
+		{"Node ID", l.sharedMonitorOptions.NodeId, "string"},
+		{"Workload ID", l.sharedMonitorOptions.WorkloadId, "string"},
+		{"Level", l.sharedMonitorOptions.Level, "string"},
 	})
 	fmt.Println(tw.Render())
 	return nil
