--- cli/monitor/validate.go
@@ -1,5 +1,0 @@
-package monitor
-
-func (m MonitorOptions) Validate() error {
-	return nil
-}
