--- cli/node/info/info.go
@@ -1,5 +1,0 @@
-package info
-
-type InfoCmd struct {
-	Id string `arg:"" required:""`
-}
