--- cli/node/info/run.go
@@ -1,5 +1,0 @@
-package info
-
-func (i InfoCmd) Run() error {
-	return nil
-}
