--- cli/node/info/table.go
@@ -1,5 +1,0 @@
-package info
-
-func (i InfoCmd) Table() error {
-	return nil
-}
