--- cli/node/info/validate.go
@@ -1,5 +1,0 @@
-package info
-
-func (i InfoCmd) Validate() error {
-	return nil
-}
