+++ cli/node/ls.go
@@ -1,0 +1,11 @@
+package node
+
+type ListCmd struct{}
+
+func (l ListCmd) Run() error {
+	return nil
+}
+
+func (l ListCmd) Validate() error {
+	return nil
+}
