--- cli/node/ls/ls.go
@@ -1,3 +1,0 @@
-package ls
-
-type ListCmd struct{}
