--- cli/node/ls/run.go
@@ -1,5 +1,0 @@
-package ls
-
-func (l ListCmd) Run() error {
-	return nil
-}
