--- cli/node/ls/table.go
@@ -1,5 +1,0 @@
-package ls
-
-func (l ListCmd) Table() error {
-	return nil
-}
