--- cli/node/ls/validate.go
@@ -1,5 +1,0 @@
-package ls
-
-func (l ListCmd) Validate() error {
-	return nil
-}
