+++ cli/node/node_up_preflight.go
@@ -2,12 +2,7 @@
 
 package node
 
-import (
-	"github.com/synadia-io/nex/cli/node/preflight"
-	"github.com/synadia-io/nex/cli/node/up"
-)
-
 type NodeExtendedCmds struct {
-	Preflight preflight.PreflightCmd `cmd:"" json:"-"`
-	Up        up.UpCmd               `cmd:"" json:"-"`
+	Preflight PreflightCmd `cmd:"" json:"-"`
+	Up        UpCmd        `cmd:"" json:"-"`
 }
