--- cli/node/preflight/table.go
@@ -1,25 +1,0 @@
-package preflight
-
-import (
-	"fmt"
-	"reflect"
-
-	"github.com/jedib0t/go-pretty/v6/table"
-	"github.com/jedib0t/go-pretty/v6/text"
-)
-
-func (p PreflightCmd) Table() error {
-	tw := table.NewWriter()
-	tw.SetStyle(table.StyleRounded)
-
-	tw.Style().Title.Align = text.AlignCenter
-	tw.Style().Format.Header = text.FormatDefault
-
-	tw.SetTitle("Node Preflight Configuration")
-	tw.AppendHeader(table.Row{"Field", "Value", "Type"})
-	tw.AppendRows([]table.Row{
-		{"Force Dependency Install", p.ForceDepInstall, reflect.TypeOf(p.ForceDepInstall).String()},
-	})
-	fmt.Println(tw.Render())
-	return nil
-}
