--- cli/node/preflight/validate.go
@@ -1,5 +1,0 @@
-package preflight
-
-func (p PreflightCmd) Validate() error {
-	return nil
-}
