+++ cli/node/table.go
@@ -1,4 +1,4 @@
-package up
+package node
 
 import (
 	"fmt"
@@ -7,6 +7,30 @@ import (
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/jedib0t/go-pretty/v6/text"
 )
+
+func (i InfoCmd) Table() error {
+	return nil
+}
+
+func (l ListCmd) Table() error {
+	return nil
+}
+
+func (p PreflightCmd) Table() error {
+	tw := table.NewWriter()
+	tw.SetStyle(table.StyleRounded)
+
+	tw.Style().Title.Align = text.AlignCenter
+	tw.Style().Format.Header = text.FormatDefault
+
+	tw.SetTitle("Node Preflight Configuration")
+	tw.AppendHeader(table.Row{"Field", "Value", "Type"})
+	tw.AppendRows([]table.Row{
+		{"Force Dependency Install", p.ForceDepInstall, reflect.TypeOf(p.ForceDepInstall).String()},
+	})
+	fmt.Println(tw.Render())
+	return nil
+}
 
 func (u UpCmd) Table() error {
 	tw := table.NewWriter()
