--- cli/node/up/validate.go
@@ -1,5 +1,0 @@
-package up
-
-func (u UpCmd) Validate() error {
-	return nil
-}
