--- cli/rootfs/validate.go
@@ -1,5 +1,0 @@
-package rootfs
-
-func (RootfsOptions) Validate() error {
-	return nil
-}
