--- cli/run/validate.go
@@ -1,9 +1,0 @@
-package run
-
-func (d DevRunOptions) Validate() error {
-	return nil
-}
-
-func (r RunOptions) Validate() error {
-	return nil
-}
