--- cli/stop/validate.go
@@ -1,5 +1,0 @@
-package stop
-
-func (s StopOptions) Validate() error {
-	return nil
-}
