--- cli/upgrade/validate.go
@@ -1,5 +1,0 @@
-package upgrade
-
-func (u UpgradeOptions) Validate() error {
-	return nil
-}
